// app/utils/Interfaces.ts
export interface AudienceItem {
  main: any;
  id: number;
  documentId: string;
  title: string;
  subtitle: string | null;
  description: string;
  icon: string; // SVG stored as string
  createdAt: string; // ISO Date
  updatedAt: string; // ISO Date
  publishedAt: string; // ISO Date
  slug: string;
}